app/requirements.txt
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.0
billiard==4.2.1
bleach==6.2.0
boto3==1.38.4
botocore==1.38.4
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
cssbeautifier==1.15.4
decorator==5.2.1
django==6.0.3
django-allauth==65.15.0
django-celery-beat==2.9.0
django-celery-results==2.6.0
django-debug-toolbar==6.2.0
django-filter==25.2
django-phonenumber-field==8.4.0
django-post-office==3.11.0
django-redis==5.4.0
django-ses==4.7.2
django-simple-history==3.11.0
django-timezone-field==7.2.1
django-widget-tweaks==1.5.0
djlint==1.36.4
editorconfig==0.17.0
executing==2.2.0
fontawesomefree==6.6.0
idna==3.10
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
pyjwt==2.10.1
kombu==5.5.3
matplotlib-inline==0.1.7
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==9.0.4
prompt-toolkit==3.0.51
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pygments==2.19.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
ruff==0.11.7
s3transfer==0.12.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.2
unidecode==1.4.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
